requirements.txt
Flask==2.3.2
gevent==23.7.0
opencv_contrib_python==4.8.0.74
opencv_python==4.8.0.74
opencv_python_headless==4.8.0.74
Pillow==9.5.0
Pillow==9.0.1
Pillow==10.0.0
python-dotenv==1.0.0
rembg==2.0.50
Requests==2.31.0
